requirements.txt
django==5.2.5
djangorestframework==3.16.1
django-cors-headers==4.7.0
gunicorn==23.0.0
psycopg2-binary==2.9.9
whitenoise==6.7.0
dj-database-url==2.3.0

